namespace restructure: glass::block contract tier + bare dispatched face, nvidia::block/warp (CUB WarpReduce L1), dispatch_body phase-1 pins, body-dispatch bench + jetson leg; attest receipt (4076/2)

+++ src/nvidia/l1.cuh
@@ -125,7 +125,7 @@ __device__ void nrm2(T *x, T *out, T *s_scratch)
     __syncthreads();
     T block_sum = BlockReduce(*reinterpret_cast<typename BlockReduce::TempStorage*>(s_scratch))
                       .Sum(thread_sum);
-    if (threadIdx.x == 0) *out = sqrtf(block_sum);
+    if (threadIdx.x == 0) *out = sqrt(block_sum);   // overload-resolved (sqrtf truncated f64)
     if constexpr (TRAILING_SYNC) {
         __syncthreads();
     }
@@ -146,3 +146,4 @@ inline constexpr std::size_t reduce_scratch_bytes()
 {
     return sizeof(typename cub::BlockReduce<T, THREADS>::TempStorage);
 }
+
